Reorganise source files and start docker work
